Redesign Homepage
Website was redesigned so that it can be scaled, without compromising on time.
Data is sourced using REST API, so users need not download the whole json file which possibly may slow down the entire site, once more data is added. Also a pro when users have slow internet.
The new design will be similar to that of authifyFame.

Co-Authored-By: authifyAlly <117624289+authifyAlly@users.noreply.github.com>