Verify by Domain
Previously by default code used to verify by hostname. So if hostname contains a sub-domain, the same sub-domain should be submitted by the user to verify that link.
With this update the code verifies wrt to the domain. So if abc.com is a verified data, any subdomain of abc.com will be automatically verified. Companies and maintainers need not update all subdomains of a domain.

Reference